# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFG API Segura - Gestión de Informes Médicos",
        "description": "API REST segura con autenticación JWT, MFA, roles y permisos, rate limiting y auditoría de seguridad.",
        "contact": {
            "name": "Pablo Sánchez",
            "email": "pablo@tfg.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos de administrador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Ver un usuario específico",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nuevo@email.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}/lock": {
            "patch": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Bloquear un usuario",
                "operationId": "00ad360a63a3dd35070ff2235dd6bd5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario bloqueado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}/unlock": {
            "patch": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Desbloquear un usuario",
                "operationId": "be8c0d92200a6a4c6dec7d00be8ed886",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario desbloqueado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/security-logs": {
            "get": {
                "tags": [
                    "Usuarios (Admin)"
                ],
                "summary": "Ver logs de seguridad",
                "operationId": "971f5f8811c0da252e04c850a20f7275",
                "responses": {
                    "200": {
                        "description": "Lista de logs de seguridad paginada"
                    },
                    "403": {
                        "description": "Sin permisos de administrador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}